Display transport summary
Upon transporting many objects or big tables you are not
able to see if something was wrong in the middle
of the transport because of the many log entries
scrolling through.
A summary is now displayed at the end of the transport
process stating any errors that occured for better
visibility.

Fixes #10