Fix delta column value comparison
The delta columns are compared between source and target systems.
The values in these columns are compared to only transport new data
to the target system.
This comparison was previously made with casts to nvarchar
on both sides.
Casting to the correct type (delta column type) removes this error.

Fixes #14